Rajout de de la prise ne compte de '\n' dans a static bufferisé et d'un ref pour la reprise de la copie du du "bluff" dans  "line"
Mise à la norme des nouvelles fonctions de la "libft"
Update des structures

+++ libft/ft_get_item.c
@@ -11,9 +11,10 @@
 /* ************************************************************************** */
 
 #include <stdlib.h>
+#include <string.h>
 #include "libft.h"
 
-t_item	*ft_get_item(t_ctrl *ctrl, int pos)
+t_item	*ft_get_item(t_ctrl *ctrl, size_t pos)
 {
 	t_item	*tmp;
